% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{AOEunits}
\alias{AOEunits}
\title{Statistics for game units in Age of Empires II: Definitive Edition}
\format{
A data frame with 128 observations of 19 variables:
\describe{
    \item{unit}{Unit name}
    \item{building}{Building in which each unit is trained}
    \item{type}{Unit class}
    \item{age}{Age at which the unit becomes trainable}
    \item{cost_wood}{Unit cost in Wood}
    \item{cost_food}{Unit cost in Food}
    \item{cost_gold}{Unit cost in Gold}
    \item{build_time}{Training time in seconds}
    \item{rate_of_fire}{Attack speed}
    \item{attack_delay}{Retasking time}
    \item{movement_speed}{Travel speed on land}
    \item{line_of_sight}{Vision over the surrounding area}
    \item{hit_points}{Unit health}
    \item{min_range}{Minimum attacking range for ranged units}
    \item{range}{Maximum attacking range for ranged units}
    \item{damage}{Damage inflicted per attack}
    \item{accuracy}{Chance that an attack will be on target}
    \item{melee_armor}{Armor against melee attacks}
    \item{pierce_armor}{Armor against projectiles}
}
}
\source{
Age of Empires II © Microsoft Corporation. This dataset was created
  under Microsoft's "Game Content Usage Rules"
  \url{https://www.xbox.com/en-us/developers/rules} using assets from Age of
  Empires II, and it is not endorsed by or affiliated with Microsoft. All
  information shown is an interpretation of data collected in-game with no
  guarantee on the accuracy of any of the data presented.
}
\usage{
AOEunits
}
\description{
A dataset with the numerical values that determine the behavior and
performance of selected military units available in AoE2:DE (July 2020 Game
Update).
}
\keyword{datasets}
