% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\encoding{UTF-8}
\name{AOEunits_raw}
\alias{AOEunits_raw}
\title{Statistics for game units in Age of Empires II: Definitive Edition in a messy presentation}
\format{
A data frame with 139 observations of 15 variables. See \code{AOEunits} for variable descriptions.
}
\source{
Age of Empires II. Copyright Microsoft Corporation. This dataset was created
under Microsoft's "Game Content Usage Rules"
\url{https://www.xbox.com/en-us/developers/rules} using assets from Age of
Empires II, and it is not endorsed by or affiliated with Microsoft. All
information shown is an interpretation of data collected in-game with no
guarantee on the accuracy of any of the data presented.
}
\usage{
AOEunits_raw
}
\description{
A messy version of the \code{\link{AOEunits}} dataset, meant for demonstrating data cleaning functions.
}
\keyword{datasets}
