% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{boutiques.xlsx}
\alias{boutiques.xlsx}
\title{boutiques.xlsx spreadsheet}
\description{
Open XML Format Spreadsheet with 1 sheet, 6 columns, and 8 rows. Toy dataset
with Q1 profits for different store locations. Additional information
is encoded as meaningful formatting. Bold indicates losses (negative values),
colors indicate continent, and italic indicates a second location in the same
city.
}
\details{
This data is used in the example for `annotate_mf_all()`.
}
