% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{dog_test.xlsx}
\alias{dog_test.xlsx}
\title{dog_test.xlsx spreadsheet}
\source{
Items are modified from the checklist written by Junior Watson.
}
\description{
Open XML Format Spreadsheet with 1 sheet, 2 columns, and 12 rows. Items
describe various tasks or behaviors that dogs can be evaluated on, assigned
into three categories which appear along with their average scores as
embedded subheaders with meaningful formatting.
}
\details{
This data is used in the example for \code{annotate_mf()}.
}
\references{
http://www.dogtrainingbasics.com/checklist-well-behaved-dog/
}
