% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{primates2017_broken}
\alias{primates2017_broken}
\title{Comparative data for 16 species of primates with some broken values}
\format{A data frame with 19 rows and 4 variables:
\describe{
    \item{scientific_name}{scientific names, with embedded subheaders
        for geographic region and taxonomic family and broken values}
    \item{common_name}{vernacular name, as listed in Estada et al.
        (2017)}
    \item{red_list_status}{IUCN Red List Status in January 2017}
    \item{mass_kg}{mean body mass in kilograms}
}}
\source{
Data originates from the appendix in the following publication:
}
\usage{
primates2017_broken
}
\description{
A dataset with embedded subheaders and some values (T. obscurus,
    T. leucocephalus and N. bengalensis) in the scientific_names variable broken up
    across two rows (typically done to fit the content in a table).
}
\references{
Estrada, Alejandro, et al. "Impending extinction crisis of the
    world’s primates: Why primates matter." Science Advances 3.1 (2017):
    e1600946. \url{http://advances.sciencemag.org/content/3/1/e1600946.full}
#' @usage
data(primates2017_broken)
}
\keyword{datasets}
