% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{primates2017_wrapped}
\alias{primates2017_wrapped}
\title{Comparative data for two species of primates}
\format{
A data frame with 9 rows and 6 variables:
\describe{
\item{scientific_name}{scientific names, see reference}
\item{common_name}{vernacular name, as listed in Estrada et al.
(2017)}
\item{habitat}{habitat types listed in the IUCN Red List assessments}
\item{red_list_status}{IUCN Red List Status in January 2017}
\item{mass_kg}{mean body mass in kilograms}
\item{country}{Countries where the species is present, from IUCN Red
List assessments}
}
}
\source{
Estrada, Alejandro, et al. "Impending extinction crisis of the
world's primates: Why primates matter." Science Advances 3.1 (2017):
e1600946. \url{https://doi.org/10.1126/sciadv.1600946}
}
\usage{
primates2017_wrapped
}
\description{
A dataset in which the elements for some of the values are in separate rows'
}
\keyword{datasets}
