% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squish_newlines.R
\name{squish_newlines}
\alias{squish_newlines}
\title{Deduplicate and remove trailing line breaks}
\usage{
squish_newlines(sepstring)
}
\arguments{
\item{sepstring}{A character vector with new line control characters.}
}
\value{
A vector without trailing or multiple consecutive new line sequences.
}
\description{
Deduplicate and remove trailing line breaks
}
\details{
Useful for tables with merged cells, often imported from Word or PDF
files. Can be applied across multiple columns before separating into rows.
}
\examples{
vecWithNewlines <- c("dog\n\ncat\n\n\npig\n")
squish_newlines(vecWithNewlines)

}
