% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unheadr-package.R
\docType{package}
\name{unheadr-package}
\alias{unheadr}
\alias{unheadr-package}
\title{unheadr: Handle Data With Embedded Subheaders}
\description{
Three verb-like functions that use tidy evaluation in dplyr and 
    tidyr to work with messy data, usually derived from spreadsheets or PDF 
    tables. Also includes functions for unwrapping values broken up across two 
    or more rows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luisDVA/unheadr}
  \item Report bugs at \url{https://github.com/luisDVA/unheadr/issues}
}

}
\author{
\strong{Maintainer}: Luis D. Verde Arregoitia \email{luis@liomys.mx} (0000-0001-9520-6543)

}
\keyword{internal}
