% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untangle_fn.R
\name{untangle2}
\alias{untangle2}
\title{Rectangling extraneous header rows.}
\usage{
untangle2(df, regex, orig, new)
}
\arguments{
\item{df}{A data frame with non-data rows embedded in the data rectangle.}

\item{regex}{Regular expression to match the extraneous headers.}

\item{orig}{Variable containing the extraneous header strings.}

\item{new}{Name of variable that will contain the extraneous header info in a
tidier format.}
}
\value{
A tibble without the extra headers and a new variable containing
  the header data.
}
\description{
Rectangling extraneous header rows.
}
\details{
For background and examples see
\url{https://luisdva.github.io/rstats/tidyeval/}. Special thanks to Jenny
Brian for fixing the initial tidyeval code and overall function structure.
}
\examples{
data(primates2017)
# put taxonomic family in its own variable (matches the suffix "DAE")
untangle2(primates2017,"DAE$",scientific_name,family)
# put geographic regions in their own variable (matching them all by name)
untangle2(primates2017,"Asia|Madagascar|Mainland Africa|Neotropics",scientific_name,family)
# with magrittr pipes
primates2017 \%>\%
    untangle2("DAE$",scientific_name,family) \%>\%
    untangle2("Asia|Madagascar|Mainland Africa|Neotropics",scientific_name,family)

}
