% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap_fn.R
\name{unwrap_cols}
\alias{unwrap_cols}
\title{Unwrap values and clean up NAs used as padding within observational units.}
\usage{
unwrap_cols(df, groupingVar, separator)
}
\arguments{
\item{df}{A data frame with wrapped values and an inconsistent number of NA
values used to as within-group padding.}

\item{groupingVar}{Name of the variable describing the observational units.}

\item{separator}{Character string defining the separator that will delimit
the elements of the unrwapped value.}
}
\value{
A summarized tibble.
}
\description{
Unwrap values and clean up NAs used as padding within observational units.
}
\details{
For more examples and background, see
\url{https://luisdva.github.io/rstats/unbreaking-vals/}.
}
\examples{
data(primates2017_wrapped)
#using commas to separate elements
unwrap_cols(primates2017_wrapped,scientific_name,", ")
}
