% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unheadr-package.R
\docType{package}
\name{unheadr-package}
\alias{unheadr}
\alias{unheadr-package}
\title{unheadr: Handle Data With Multiple Header Rows}
\description{
Three verb-like functions that use tidy evaluation in dplyr and
tidyr to work with messy data, usually derived from spreadsheets or pdf
tables. Also includes functions for unwrapping values broken up across two
or more rows.
}
\author{
\strong{Maintainer}: Luis D. Verde Arregoitia \email{luis@liomys.mx}

}
\keyword{internal}
