% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unheadr-package.R
\docType{package}
\name{unheadr-package}
\alias{unheadr}
\alias{unheadr-package}
\title{unheadr: Handle Data with Messy Header Rows and Broken Values}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Verb-like functions to work with messy data, often derived from 
             spreadsheets or parsed PDF tables. Includes functions for unwrapping 
             values broken up across rows, relocating embedded grouping values, 
             and to annotate meaningful formatting in spreadsheet files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luisDVA/unheadr}
  \item Report bugs at \url{https://github.com/luisDVA/unheadr/issues}
}

}
\author{
\strong{Maintainer}: Luis D. Verde Arregoitia \email{luis@liomys.mx} (\href{https://orcid.org/0000-0001-9520-6543}{ORCID})

}
\keyword{internal}
